Refactor some low-level functions into a helper module.
